requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
bitarray==1.2.2
blinker==1.4
Brlapi==0.7.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
cytoolz==0.11.2
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-filter==22.1
django-metamask-auth==0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
dnspython==1.16.0
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fasteners==0.14.1
filelock==3.0.12
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
hexbytes==0.2.2
httplib2==0.14.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.12.0
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.7
MarkupSafe==2.1.1
monotonic==1.5
more-itertools==4.2.0
multiaddr==0.0.9
multidict==6.0.2
mysqlclient==2.1.1
netaddr==0.8.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
parsimonious==0.8.1
pbr==5.11.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.20.1
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.14.0
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.15.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
pythondialog==3.4.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
rlp==2.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
toolz==0.11.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.10
usb-creator==0.3.7
varint==1.0.2
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
web3==5.28.0
websockets==9.1
whitenoise==6.4.0
xkit==0.0.0
yarl==1.7.2
zipp==3.8.0
